Changed data-type of nParents, nChildren of Struct skill & nUpdates, nPreviousUpdates of Struct ReputationLogEntry from uint256 to uint128  of Struct Skill, Updated the tests to support one storage slot instead of two in file colony-network-recovery.js .
Fixed the tests, Everything should work fine now.

Changed data-type of nParents, nChildren of Struct skill & nUpdates, nPreviousUpdates of Struct ReputationLogEntry from uint256 to uint128 of Struct Skill, Updated the tests to support one storage slot instead of two in file colony-network-recovery.js

Reverted the type conversion

+++ contracts/ColonyNetworkDataTypes.sol
@@ -97,9 +97,9 @@ contract ColonyNetworkDataTypes {
 
   struct Skill {
     // total number of parent skills
-    uint256 nParents;
+    uint128 nParents;
     // total number of child skills
-    uint256 nChildren;
+    uint128 nChildren;
     // array of `skill_id`s of parent skills starting from the 1st to `n`th, where `n` is an integer power of two larger than or equal to 1
     uint256[] parents;
     // array of `skill_id`s of all child skills
@@ -118,7 +118,7 @@ contract ColonyNetworkDataTypes {
     int amount;
     uint256 skillId;
     address colony;
-    uint256 nUpdates;
-    uint256 nPreviousUpdates;
+    uint128 nUpdates;
+    uint128 nPreviousUpdates;
   }
 }
